==> notebooks/01_LoadData.ipynb <==
### Load Data from URL

#### Output
* winequality-red.csv

### Azure Blob Storage

import os
from azure.storage.blob import BlockBlobService

# azure credentials
# make sure the credentials are saved as environment variables.
os.environ["AZURE_STORAGE_CONNECTION_STRING"] = 'xxx'
connect_str = os.getenv('AZURE_STORAGE_CONNECTION_STRING')
# print(connect_str)

# file and directory info
container_name = 'datascience-data' 
file_name = 'winequality-red.csv'
project_dir = os.path.dirname(os.getcwd())
final_file_dir = os.path.join(project_dir, 'data/raw', file_name)

# for azure.storage.blob 2.1.0
blob = BlockBlobService(connection_string=connect_str)
blob.get_blob_to_path(container_name, 
                      file_name, 
                      final_file_dir)

### URL

# import os
# import pandas as pd

# csv_url = 'https://raw.githubusercontent.com/choux130/Wine-Quality-Dataset/master/winequality-red.csv'
# data = pd.read_csv(csv_url, error_bad_lines = False)

# # file and directory info
# file_name = 'winequality-red.csv'
# project_dir = os.path.dirname(os.getcwd())

# # save the file to the assigned directory
# final_file_dir = os.path.join(project_dir, 'data/raw', file_name)
# data.to_csv(final_file_dir, index=None, header=True, float_format='%g')
