==> notebooks/02_SplitData.ipynb <==
### Split Data into Train and Test dataset

#### Input 
* winequality-red.csv

#### Output
* train.parquet
* test.parquet

# params 
random_seed = 123
percent_train = 0.8

import os
import pandas as pd 
import pyarrow
from sklearn.model_selection import train_test_split

# file and directory info
project_dir = os.path.dirname(os.getcwd())
interim_folder = "/data/interim/"
raw_folder = "/data/raw/"
file_name = 'winequality-red.csv'
var_y = 'quality'

data = pd.read_csv(project_dir + raw_folder + file_name, error_bad_lines = False)
train, test = train_test_split(data, train_size = 0.2, random_state = random_seed)

train.to_parquet(project_dir + interim_folder + 'train.parquet')
test.to_parquet(project_dir + interim_folder + 'test.parquet')

print(train.shape)
train.head(10)

print(test.shape)
train.head(10)
