==> notebooks/06_Predict.ipynb <==
### Predict using API

##### In terminal ===================================================================================================
# # train a model
# !! python ../code/a05_Train.py 
# # or, but have to create the base image "ds_project_template" first
!! mlflow run ../../DSProjectTemplate_Python

### Host locally

##### In terminal ===================================================================================================
# # serve the model locally with port 7000, the path for the model can be in Azure blob storage or local path
# don't run in Notebook since it will never end.
# mlflow models serve -m < path to the > -p 7000

# # get prediction using POST
!! curl http://127.0.0.1:7000/invocations -H "Content-Type:application/json; format=pandas-split" --data '{"columns":["alcohol", "chlorides", "citric acid", "density", "fixed acidity", "free sulfur dioxide", "pH", "residual sugar", "sulphates", "total sulfur dioxide", "volatile acidity"],"data":[[1,1, 1, 1, 1, 1, 1, 1, 1, 1, 1]]}'

### Host in Azure ML

##### In terminal ===================================================================================================
# # create image in the Azure ML
# mlflow azureml build-image -w <workspace-name> -m <model-path> -d "Wine regression model 1"

##### In terminal ===================================================================================================
# # this command is deprecated in the most recent version. but we can go to Azure protal and deploy the image by clicking
# az ml service create aci -n <deployment-name> --image-id <image-name>:<image-version>

# # get prediction using POST
# don't run in company internet or wifi, it will be blocked
!! curl <Scoring URI> -H "Content-Type:application/json; format=pandas-split" --data '{"columns":["alcohol", "chlorides", "citric acid", "density", "fixed acidity", "free sulfur dioxide", "pH", "residual sugar", "sulphates", "total sulfur dioxide", "volatile acidity"],"data":[[1,1, 1, 1, 1, 1, 1, 1, 1, 1, 1]]}'
